Hpc event

+++ RunJobHpcEvent.py
@@ -1286,7 +1286,7 @@ class RunJobHpcEvent(RunJob):
                     self.__jobs[jobId]['job'].timeStageOut = 0
                     self.__jobs[jobId]['job'].nEvents = 0
                     self.__jobs[jobId]['job'].nEventsW = 0
-                    self.__jobs[jobId]['job'].yodaJobMetrics = None
+                    self.__jobs[jobId]['job'].yodaJobMetrics = {'startTime': time.time(), 'endTime': time.time()}
                     job = self.__jobs[jobId]['job']
                     job.cpuConversionFactor = 1
                     tolog("Job CPU usage: %s %s" % (job.cpuConsumptionTime, job.cpuConsumptionUnit))
@@ -1309,6 +1309,9 @@ class RunJobHpcEvent(RunJob):
 
     def zipOutputs(self, job, zipEventRangeName, zipFileName):
         eventstatus = str(job.jobId) + "_event_status.dump"
+        if os.path.exists(eventstatus + ".zipped"):
+            tolog("Event status dump file %s exist. It's already zipped." % eventstatus + ".zipped")
+            return
         if not os.path.exists(eventstatus):
             tolog("Event status dump file %s doesn't exist. checking backup file" % eventstatus)
             eventstatus = eventstatus + ".backup"
@@ -1341,11 +1344,16 @@ class RunJobHpcEvent(RunJob):
             outputs = output.split(",")[:-3]
             for out in outputs:
                 #tolog("Adding file: %s" % out)
+                if not os.path.exists(out):
+                    tolog("File %s doesn't exist" % out)
+                    continue
                 tar.add(out, arcname=os.path.basename(out))
                 os.remove(out)
             zipEventRange.write("%s %s %s\n" % (eventRangeID, status, output))
         tar.close()
         zipEventRange.close()
+        tolog("Zip finished, Rename %s to %s" % (eventstatus, eventstatus + ".zipped"))
+        os.rename(eventstatus, eventstatus + ".zipped")
 
     def stageOutZipFile(self, job, espath, os_bucket_id):
         try:
@@ -1367,10 +1375,13 @@ class RunJobHpcEvent(RunJob):
 
             if self.__copyOutputToGlobal:
                 outputDir = os.path.dirname(os.path.dirname(zipFileName))
-                tolog("Copying tar/zip file %s to %s" % (zipFileName, os.path.join(outputDir, os.path.basename(zipFileName))))
+                tolog("Moving tar/zip file %s to %s" % (zipFileName, os.path.join(outputDir, os.path.basename(zipFileName))))
                 os.rename(zipFileName, os.path.join(outputDir, os.path.basename(zipFileName)))
                 tolog("Copying tar/zip file %s to %s" % (zipEventRangeName, os.path.join(outputDir, os.path.basename(zipEventRangeName))))
-                os.rename(zipEventRangeName, os.path.join(outputDir, os.path.basename(zipEventRangeName)))
+                shutil.copyfile(zipEventRangeName, os.path.join(outputDir, os.path.basename(zipEventRangeName)))
+                eventstatusFile = str(job.jobId) + "_event_status.dump"
+                tolog("Copying dump file %s to %s" % (eventstatusFile, os.path.join(outputDir, os.path.basename(eventstatusFile))))
+                shutil.copyfile(eventstatusFile, os.path.join(outputDir, os.path.basename(eventstatusFile)))
                 return
 
             report = getInitialTracingReport(userid=job.prodUserID, sitename=self.__jobSite.sitename, dsname=dsname, eventType="objectstore", analysisJob=False, jobId=job.jobId, jobDefId=job.jobDefinitionID, dn=job.prodUserID)
@@ -1385,7 +1396,8 @@ class RunJobHpcEvent(RunJob):
                     line = line.strip()
                     if len(line):
                         eventRangeID = line.split(" ")[0]
-                        eventRanges.append({'eventRangeID': eventRangeID, 'eventStatus': 'finished', 'objstoreID': os_bucket_id})
+                        eventStatus = line.split(" ")[1]
+                        eventRanges.append({'eventRangeID': eventRangeID, 'eventStatus': eventStatus, 'objstoreID': os_bucket_id})
                 if job.yodaJobMetrics and 'totalProcessedEvents' in job.yodaJobMetrics:
                     job.yodaJobMetrics['totalProcessedEvents'] = len(eventRanges)
                     job.nEventsW = job.yodaJobMetrics['totalProcessedEvents']
@@ -1576,8 +1588,8 @@ class RunJobHpcEvent(RunJob):
 
         tolog("The job state is %s" % _job.jobState)
         if _job.jobState in ['starting', 'transfering']:
-            tolog("The job hasn't started to run, will not recover it. Just finish it.")
-            return False
+            tolog("The job hasn't started to run")
+            # return False
 
         os.chdir(self.__jobSite.workdir)
         self.__jobs[_job.jobId]['JR'] = JobRecovery(pshttpurl='https://pandaserver.cern.ch', pilot_initdir=_job.workdir)
@@ -1643,6 +1655,9 @@ class RunJobHpcEvent(RunJob):
                     # return -1
                     pilotErrorDiag="Not All events are handled(total:%s, left:%s)" % (job.nEvents, job.nEventsW)
 
+                # Panda only record nEvents
+                job.nEvents = job.nEventsW
+
         dsname, datasetDict = self.getJobDatasets(job)
         tolog("dsname = %s" % (dsname))
         tolog("datasetDict = %s" % (datasetDict))
@@ -1670,15 +1685,15 @@ class RunJobHpcEvent(RunJob):
         self.moveTrfMetadata(job.workdir, job.jobId)
 
         # Check the job report for any exit code that should replace the res_tuple[0]
-        res0, exitAcronym, exitMsg = self.getTrfExitInfo(0, job.workdir)
-        res = (res0, exitMsg, exitMsg)
+        # res0, exitAcronym, exitMsg = self.getTrfExitInfo(0, job.workdir)
+        # res = (res0, exitMsg, exitMsg)
 
         # Payload error handling
-        ed = ErrorDiagnosis()
-        job = ed.interpretPayload(job, res, False, 0, self.__jobs[job.jobId]['runCommandList'], self.getFailureCode())
-        if job.result[1] != 0 or job.result[2] != 0:
-            self.failOneJob(job.result[1], job.result[2], job, pilotErrorDiag=job.pilotErrorDiag, final=True, updatePanda=False)
-            return -1
+        # ed = ErrorDiagnosis()
+        # job = ed.interpretPayload(job, res, False, 0, self.__jobs[job.jobId]['runCommandList'], self.getFailureCode())
+        # if job.result[1] != 0 or job.result[2] != 0:
+        #     self.failOneJob(job.result[1], job.result[2], job, pilotErrorDiag=job.pilotErrorDiag, final=True, updatePanda=False)
+        #     return -1
 
         self.updateJobState(job, "finished", "finished", final=False)
 
@@ -1726,7 +1741,7 @@ class RunJobHpcEvent(RunJob):
                 try:
                     if "job.log.tgz." in file or "LOCKFILE" in file or "tarball_PandaJob" in file or "objectstore_info" in file:
                         continue
-                    if file.endswith(".dump") or file.startswith("metadata-") or "jobState-" in file\
+                    if file.endswith(".dump") or file.endswith(".dump.zipped") or file.startswith("metadata-") or "jobState-" in file\
                        or file.startswith("jobState-") or file.startswith("EventService_premerge")\
                        or file.startswith("Job_") or file.startswith("fileState-") or file.startswith("curl_updateJob_")\
                        or file.startswith("curl_updateEventRanges_")\
@@ -1744,6 +1759,16 @@ class RunJobHpcEvent(RunJob):
         except:
             tolog(sys.exc_info()[1])
             tolog(sys.exc_info()[2])
+
+        try:
+            self.checkJobMetrics()
+        except:
+            tolog("RunHPCEvent failed: %s" % traceback.format_exc())
+
+        try:
+            self.stageOutZipFiles()
+        except:
+            tolog("RunHPCEvent failed: %s" % traceback.format_exc())
 
         try:
             tolog("Copying Log files to Job working dir")
